Update message oriented read write spec
Rather than supporting partial read/write of vmsg objects, it makes
more sense to make these write opaque.  That means we'll need some
additioanl functionality on the vmsg structs and some functions to
handle resetting/preparing the accounting values in the vmsg structs.

The upside is that operations on vmsg will be simpler for the caller.