Fixed issue in PVM driver where orphaned sockets were not closed.
Special care has been taken to handle the rundown case. In particular we must be aware whether a request to destroy a socket came from user mode (our program) or from kernel mode (e.g the orphaned socket). Moreover, we must ensure that each thread-response mapping is associated with only one in-flight request/response at a time. We must forceably close a socket when the thread has nothing in-flight.