Fixed driver problem, updated some types in messages.
Driver was not handling rundown correctly when a server was using it.  For instance, the server would call accept() and upon rundown, the driver would wait forever for the response to the accept() request to arrive.

Also, the Rump xenevent_app code was not handling a failed accept() correctly. Even when accept failed, it would allocate a new thread for it, because the failure was not recorded correctly.

Message types have been updated so we can easily ID messages that allocate or deallocate a FD. Also, the status field was reduced back down to 32 bits.