This commit adds a break to a switch statement in translate.h which fixes a bug where an invalid socket type message was being printed out erroniously.
This commit adds the apache bench performance tests to the $MWROOT/test/apache_bench
directory in the form of two scripts.  make_files.sh and run_benchmark.sh.

The two scripts must both be run to work and the detials are in the README file in that directory.

Code on brothour.