Candidate commit to fix performance problems.
In an earlier attempt to address performance issues (commit 67075c) a design
flaw was introduced, which is fixed here.

For maximum performance, the INS side of the comms channel should be
event-driven, meaning that the PVM side updates the ring buffer and
then sends an event on the shared channel. However, the aforementioned
commit altered the PVM side to presume that the INS would actively monitor
the ring buffer itself and pull requests off as they became available,
independent of whether the event channel had been signalled. This erroneous
commit caused the INS not to recognize a new request until there was a
backlog of several requests, thus slowing down responsiveness.