Stubbed out mitigation message support.
Renamed backchannel to netflow; fixed several issues including:
* Incorrect handling of netflow socket when there was nothing to read
  caused deadlock.
* Sample netflow consumer had many bugs

No mitigation processing is done, but the general framework is in place.
Next steps on the PVM side involve converting a mitgation request into
an action, either by sending a request to the INS or by doing an action
on the PVM (like killing a thread).